+++ R/mcmc-diagnostics.R
@@ -18,6 +18,13 @@
 #'
 #' @section Plot Descriptions:
 #' \describe{
+#' \item{`mcmc_rhat_data()`, `mcmc_neff_data()`}{
+#'   Data-preparation back ends for the R-hat and effective sample size plots.
+#'   Users can call these functions directly to obtain the data frame of
+#'   diagnostic values with rating labels and create custom diagnostic
+#'   visualizations with **ggplot2**.
+#' }
+#'
 #' \item{`mcmc_rhat()`, `mcmc_rhat_hist()`}{
 #'   Rhat values as either points or a histogram. Values are colored using
 #'   different shades (lighter is better). The chosen thresholds are somewhat
@@ -70,7 +77,7 @@
 #' (p <- mcmc_acf_bar(x, pars = c("alpha", "beta[1]")))
 #'
 #' # add horiztonal dashed line at 0.5
-#' p + hline_at(0.5, linetype = 2, size = 0.15, color = "gray")
+#' p + hline_at(0.5, linetype = 2, linewidth = 0.15, color = "gray")
 #' }
 #'
 #' # fake rhat values to use for demonstration
@@ -248,8 +255,8 @@ mcmc_neff <- function(ratio, ..., size = NULL) {
       linetype = 2,
       linewidth = 0.25) +
     labs(y = NULL, x = expression(N[eff]/N)) +
-    scale_fill_diagnostic("neff") +
-    scale_color_diagnostic("neff") +
+    scale_fill_diagnostic("neff_ratio") +
+    scale_color_diagnostic("neff_ratio") +
     scale_x_continuous(
       breaks = breaks,
       # as.character truncates trailing zeroes, while ggplot default does not
@@ -280,8 +287,8 @@ mcmc_neff_hist <- function(ratio, ..., binwidth = NULL, bins = NULL, breaks = NU
       binwidth = binwidth,
       bins = bins,
       breaks = breaks) +
-    scale_color_diagnostic("neff") +
-    scale_fill_diagnostic("neff") +
+    scale_color_diagnostic("neff_ratio") +
+    scale_fill_diagnostic("neff_ratio") +
     labs(x = expression(N[eff]/N), y = NULL) +
     dont_expand_y_axis(c(0.005, 0)) +
     yaxis_title(FALSE) +
@@ -419,12 +426,12 @@ diagnostic_points <- function(size = NULL) {
 
 # Functions wrapping around scale_color_manual() and scale_fill_manual(), used to
 # color the intervals by rhat value
-scale_color_diagnostic <- function(diagnostic = c("rhat", "neff")) {
+scale_color_diagnostic <- function(diagnostic = c("rhat", "neff_ratio")) {
   d <- match.arg(diagnostic)
   diagnostic_color_scale(d, aesthetic = "color")
 }
 
-scale_fill_diagnostic <- function(diagnostic = c("rhat", "neff")) {
+scale_fill_diagnostic <- function(diagnostic = c("rhat", "neff_ratio")) {
   d <- match.arg(diagnostic)
   diagnostic_color_scale(d, aesthetic = "fill")
 }