+++ R/ppc-test-statistics.R
@@ -55,6 +55,13 @@
 #'    computed over the datasets (rows) in `yrep`. The value of the
 #'    statistics in the observed data is overlaid as large point.
 #'   }
+#'   \item{`ppc_stat_data()`}{
+#'    Data-preparation back end for `ppc_stat()`, `ppc_stat_freqpoly()`, and
+#'    their grouped variants. Users can call `ppc_stat_data()` directly to
+#'    obtain the data frame of test-statistic values computed from `y` and
+#'    each row of `yrep`, enabling custom test-statistic visualizations with
+#'    **ggplot2**.
+#'   }
 #' }
 #'
 #' @examples
@@ -442,9 +449,9 @@ Tyrep_label <- function() expression(italic(T)(italic(y)[rep]))
 
 message_if_using_mean <- function(stat) {
   if (is.character(stat) && stat == "mean") {
-    message(
+    inform(paste0(
       "Note: in most cases the default test statistic 'mean' is ",
       "too weak to detect anything of interest."
-    )
+    ))
   }
 }